Move a couple files around; WebhookErrorTest will not modify the infrastructure shortly
